Added support for an actor messaging itself.
Removed the previous hotactor support because it didn't handle an actor's action message queue correctly.  The new method is a bit slower than the previous one (0.363 vs 0.330 for threadring).  It may be worth revisiting in the future for that reason, but its run loop had become so similar to the main one that it didn't really seem like a good idea to keep it around if we could move away from it.