Added the ability to exit an action prematurely with 'return' (which will wait for more input) or 'done' (which will close and delete that actor).
Also fixed a problem where timeslices were not being decremented correctly, causing the maintenance loop to not run in severe cases.  This loses some performance on threadring again (from ~0.36 to ~0.39).  This is to be expected, as the optimizations being folded in are for a different class of problem than threadring.